adds sign out functionality
