+++ app/lib/user.rb
@@ -29,5 +29,4 @@ class User
       return nil
     end
   end
-
 end
