+++ app/views/layout.erb
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Awesome Twitter</title>
+    <title>Chitter</title>
     <link rel="stylesheet" type="text/css" href="/css/application.css">
   </head>
   <body>
@@ -10,12 +10,25 @@
     <% end %>
     <% if flash[:errors] && !flash[:errors].empty? %>
       Sorry, there were the follow problems with the form:
-        <ul id='errors'>
-          <% flash[:errors].each do |error| %>
-            <li><%= error %></li>
-          <% end %>
-        </ul>
+      <ul id='errors'>
+        <% flash[:errors].each do |error| %>
+        <li><%= error %></li>
+        <% end %>
+      </ul>
     <% end %>
+
+    <% if current_user %>
+      <form method='post' action='/sessions'>
+        <input type='hidden' name='_method' value='delete'>
+        <input type='submit' value='Sign out'>
+      </form>
+      <% if flash[:notice] %>
+        <div id="notice">
+          <%= flash[:notice] %>
+        </div>
+      <% end %>
+    <% end %>
+
     <%= yield %>
   </body>
 </html>
