+++ app/views/sessions/new.erb
@@ -1,4 +1,4 @@
-Please sign in.
+<h1>Sign in</h1>
 
 <form method='post' action='/sessions'>
   Email: <input type='text' name='email'>
