+++ app/views/users/new.erb
@@ -1,4 +1,4 @@
-<h1>Please sign up</h1>
+<h1>Sign up</h1>
 
 <form action='/users' method='post'>
   Name: <input name='name' type='text'>
