+++ spec/features/user_management_spec.rb
@@ -1,4 +1,7 @@
 require 'spec_helper'
+require './app/helpers/sessions'
+
+include SessionHelpers
 
 feature 'User signs up' do
 
@@ -18,19 +21,6 @@ feature 'User signs up' do
     expect { sign_up }.to change(User, :count).by(1)
     expect { sign_up }.to change(User, :count).by(0)
     expect(page).to have_content("This email is already taken")
-  end
-
-  def sign_up(email = 'sammy@example.com', password = 'lolol',
-              password_confirmation = 'lolol', name = 'Sammy',
-              username = 'Vinyls')
-    visit '/users/new'
-    expect(page.status_code).to eq(200)
-    fill_in :name, with: name
-    fill_in :username, with: username
-    fill_in :email, with: email
-    fill_in :password, with: password
-    fill_in :password_confirmation, with: password_confirmation
-    click_button 'Sign up'
   end
 end
 
@@ -52,11 +42,19 @@ feature 'User signs in' do
     sign_in('test@test.com', 'wrong')
     expect(page).not_to have_content('Welcome, Dude!')
   end
+end
 
-  def sign_in(email, password)
-    visit '/sessions/new'
-    fill_in 'email', with: email
-    fill_in 'password', with: password
-    click_button 'Sign in'
+feature 'User signs out' do
+  before(:each) do
+    User.create(email: 'test@test.com',
+                password: 'test',
+                password_confirmation: 'test')
   end
+
+  # scenario 'while being signed in' do
+  #   sign_in('test@test.com', 'test')
+  #   click_button 'Sign out'
+  #   expect(page).to have_content('Cya next time!')
+  #   expect(page).not_to have_content('Welcome, test@test.com')
+  # end
 end
